fix unsigned hex literals

+++ vlib/crypto/sha512/sha512block_generic.v
@@ -13,7 +13,7 @@ import math.bits
 
 const(
 	_K = [
-		0x428a2f98d728ae22,
+		u64(0x428a2f98d728ae22),
 		0x7137449123ef65cd,
 		0xb5c0fbcfec4d3b2f,
 		0xe9b5dba58189dbbc,
